FastApiAzureDB/requirements.txt
pyodbc~=5.0.1
SQLAlchemy~=1.4.49
uvicorn~=0.23.2
fastapi~=0.104.0
starlette~=0.27.0
pandas~=2.1.1
PyJWT~=2.8.0
pydantic~=2.4.2
